requirements.txt
folium
streamlit
streamlit_folium
seaborn
pandas
pymongo
matplotlib
wordcloud
plotly
numpy
requests
spacy
geopy
firebase-admin
python-dotenv
